projects/mythological-forge/requirements.txt
# The Mythological Forge - Python Requirements

# Core dependencies
pyyaml>=6.0
jsonschema>=4.0
pathlib>=1.0

# HTML processing
beautifulsoup4>=4.12.0
html2text>=2020.1.16

# Command line
argparse>=1.4

# Optional: Database support
# sqlite3 (included in Python standard library)

# Optional: Enhanced search
# elasticsearch>=8.0
# whoosh>=2.7

# Optional: Web interface
# flask>=3.0
# flask-cors>=4.0

# Optional: NLP for content analysis
# spacy>=3.0
# nltk>=3.8

# Development tools
# pytest>=7.0
# black>=23.0
# flake8>=6.0